page遷移とdelete時の消去バグを解消

+++ app/features/db/table.ts
@@ -5,7 +5,7 @@ import { SupabaseClient } from "@supabase/supabase-js";
 import { Database } from "@/app/types/db_types"; 
 import { error } from "console";
 import { create } from "domain";
-import { deleteImage } from "@/app/features/db/bucket";
+import { deleteImage } from "@/app/features/db/bucket_sa";
 //要型注釈
 
 //全wordとそれと紐づくsentenceを取りだして返す
@@ -18,17 +18,18 @@ export const getData=async():Promise<{words:Word[],sentences:Sentence[]}|null>=>
     .order("id",{ascending:true})
     let words:Word[]=[]
     let sentences:Sentence[]=[]
-    if(data){
-        data.map((element)=>{
-            if(element){
-                words.push(element)
-                if(element.sentence_list){
-                    sentences.push(element.sentence_list)
-                    console.log(element.sentence_list)
-                }
-            } 
-        })
-    }
+    if(!data)throw new Error("データ取得に失敗しました")
+    data.map((element)=>{
+        if(element){
+            words.push(element)
+            if(element.sentence_list){
+                sentences.push(element.sentence_list)
+                console.log(element.sentence_list)
+            }
+        } 
+        
+    })
+    if(words.length===0)throw new Error("データがありません")
     return {"words":words,"sentences":sentences}
     }
     catch(error){
@@ -78,6 +79,7 @@ export const insertSentence=async(sentence:Pick<Sentence,"id"|"sentence"|"senten
     .select()
     .single()
     if(data){
+        console.log(data)
         return data
     }else{
         throw new Error("データ取得に失敗しました")
@@ -89,7 +91,7 @@ export const insertSentence=async(sentence:Pick<Sentence,"id"|"sentence"|"senten
 }
 
 //IDを指定して削除(このとき画像も削除する)
-export const delWordData=async(ID:number,wordImage:string,sentenceImage:string|null):Promise<boolean>=>{
+export const delData=async(ID:number,wordImage:string|null,sentenceImage:string|null):Promise<boolean>=>{
     try{
         const supabase=await createClient()
         //画像削除
@@ -99,16 +101,32 @@ export const delWordData=async(ID:number,wordImage:string,sentenceImage:string|n
             if(!deleteSentenceImage){
                 throw new Error("例文画像の削除に失敗しました")
             }
+            //例文データの削除
+            const {error}=await supabase
+            .from("sentence_list")
+            .delete()
+            .eq("id",ID)
+            .select()
+            if(error){
+                throw new Error("例文データの削除に失敗しました")
+            }
+            
         }
-        const deleteWordImage=await deleteImage(wordImage)
-
-        if(!deleteWordImage){
-            throw new Error("画像の削除に失敗しました")
-        }else{
-            await supabase
+        if(wordImage){
+            const deleteWordImage=await deleteImage(wordImage)
+            if(!deleteWordImage){
+                throw new Error("単語画像の削除に失敗しました")
+            }
+            //単語の削除
+            const {error}=await supabase
             .from("word_list")
             .delete()
             .eq("id",ID)
+            .select()
+            if(error){
+                throw new Error("単語データの削除に失敗しました")
+            }
+
         }
         return true
